Change log format order to time, module, level, message
Before the log format was:
    "%(asctime)s - %(levelname)s - %(name)s - %(message)s"
And now it is:
    "%(asctime)s - %(name)s - %(levelname)s - %(message)s"

Reason for this is purely cosmetic. Since several log messages in a row
will have the same value for 'name' it looks better to have the name
before the level.

Example lines from old format:
    2013-10-14 20:24:13,423 - DEBUG - root - debug message here
    2013-10-14 20:24:13,589 - WARNING - root - warning message here

And from new format:
    2013-10-14 20:24:13,423 - root - DEBUG - debug message here
    2013-10-14 20:24:13,589 - root - WARNING - warning message here

The new format will tend to keep more things lined up.