Specify coding in a more popular format
Previously it was in a vim modeline format.  Changed it to the emacs
format since that's far and away the more popular way of specifying
utf-8.  Plus vim had a lot of rules around when it would actual read
that setting and a lot of vim installs now disable modelines for
security.